add new features

+++ src/main/java/com/example/bankaccounts/repository/BankAccountRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BankAccountRepository extends JpaRepository<BankAccount,Long> {
+public interface BankAccountRepository extends JpaRepository<BankAccount,Integer> {
 }
